* p4a-pips:
  Add unfolding, pyps made ! Just great :)
  Don't disable warning on hash_put globally, instead use overwrite_ordering_of_the_statement_to_current_mapping() from time to time
  Add overwrite_ordering_of_the_statement_to_current_mapping()
  Don't warn on hash_put redefinition in loop fusion
  Add Makefile.in to svn:ignore property
  Casting from pointer to integer (bool) is forbidden, use intptr_t instead
  Revert 20186: casting bool as integer

+++ packages/PIPS/pips/src/Libs/ri-util/entity.c
@@ -1553,7 +1553,7 @@ static bool comparable_entity_in_list_p(entity common, entity v, list l)
 
   pips_debug(4, "%s ~ %s? %d: n=%d,o=%d,t=%d\n", entity_name(v),
 	     entity_undefined_p(ref)? "<undef>": entity_name(ref),
-	                            (int)ok, (int)sn, (int)so, (int)st);
+	     ok, sn, so, st);
 
   /* temporary for CA
    */
@@ -1561,7 +1561,7 @@ static bool comparable_entity_in_list_p(entity common, entity v, list l)
     pips_debug(1, "common /%s/: %s != %s (n=%d,o=%d,t=%d)\n",
 	       entity_name(common), entity_name(v),
 	       entity_undefined_p(ref)? "<undef>": entity_name(ref),
-	                              (int)sn, (int)so, (int)st);
+	       sn, so, st);
   }
 
   return ok;
